requirements.txt
absl-py==0.12.0
asgiref==3.3.4
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
Cython==0.29.22
dataclasses==0.6
decorator==4.4.2
Django==3.2.4
future==0.18.2
google-auth==1.28.0
google-auth-oauthlib==0.4.3
grpcio==1.36.1
idna==2.10
imageio==2.9.0
importlib-metadata==3.7.3
joblib==1.0.1
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
networkx==2.5.1
numpy==1.20.1
oauthlib==3.1.0
opencv-python==4.5.1.48
pandas==1.1.5
Pillow==8.1.2
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
sqlparse==0.4.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.2
threadpoolctl==2.1.0
tifffile==2021.4.8
torch==1.7.0
torchvision==0.8.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
xlrd==1.2.0
zipp==3.4.1
